Do not backup symlinks

+++ scripts/symlinks.sh
@@ -6,6 +6,8 @@ find "$HOME/.dotfiles/home" -type d | sed 's/\.dotfiles\/home\///' | xargs mkdir
 # create symlinks for all files, backup old files
 while read src; do
   target=$(echo $src | sed 's/\.dotfiles\/home\///')
-  mv "$target" "$target.bak"
+  if [[ ! -L "$target" && -f "$target" ]]; then
+    mv "$target" "$target.bak"
+  fi
   ln -sf "$src" "$target"
 done < <(exec find "$HOME/.dotfiles/home" -type f)
